% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_business_days.R
\name{count_business_days}
\alias{count_business_days}
\title{Calculating the Number of Business Days}
\usage{
count_business_days(year, from_month, to_month)
}
\arguments{
\item{year}{target year}

\item{from_month}{start of the range}

\item{to_month}{end of the range}
}
\description{
Calculating the Number of Business Days
}
