% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input summary.R
\name{inputSummary}
\alias{inputSummary}
\title{Provide summary of input variables}
\usage{
inputSummary()
}
\description{
Provides input variable information potentially useful to front-end developer of online calculator tool.
}
