% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal rate.R
\name{margRate}
\alias{margRate}
\title{Estimate household marginal tax rate}
\usage{
margRate(d)
}
\description{
Approximaton of household marginal federal income tax rate using tax year
2017 values. Assumes gross household income is equivalent to AGI and applies
standard deduction in all cases. Households with >2 adults are treated as
married with all adults >2 counted as dependents.
}
