% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict model.R
\name{predictModel}
\alias{predictModel}
\title{Use prediction models to convert user-provided inputs into calculator return values}
\usage{
predictModel(input)
}
\description{
Use prediction models to convert user-provided inputs into calculator return values
}
