% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict model.R
\name{predictModel}
\alias{predictModel}
\title{Predict a simple GBM model (test)}
\usage{
predictModel(input)
}
\arguments{
\item{input}{data passed on as \code{newdata} to \code{\link{predict.gbm}}}
}
\description{
Simple model with two predictor \code{X1} and \code{X2}
}
\examples{
mydata <- data.frame(
   X1=runif(3),
   X2=2*runif(3)
)
predictModel(mydata)
}
