% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict model.R
\name{predictModel}
\alias{predictModel}
\title{Predict a simple GBM model}
\usage{
predictModel(input)
}
\arguments{
\item{input}{data passed on as \code{newdata} to \code{\link{predict.gbm}}}
}
\description{
Simple model with multiple predictors
}
